chore: move hooks to internalmanager folder
