+++ packages/next-app/src/hooks/internalmanager/useTokenApproval.ts
@@ -3,8 +3,8 @@ import { Chain, waitForTransaction } from "@wagmi/core";
 import { useInternalBalance } from "#/contexts/InternalManagerContext";
 import { writeTokenApproval } from "#/wagmi/writeTokenApproval";
 
+import { SubmitData, TransactionStatus } from "../useTransaction";
 import { NOTIFICATION_MAP_INTERNAL_BALANCES } from "./useManageUserBalance";
-import { SubmitData, TransactionStatus } from "./useTransaction";
 import { useTransactionStatus } from "./useTransactionStatus";
 
 export function useTokenApproval() {
